code optimizations and error messages fix
